requirements.txt
Flask==2.3.3
requests==2.31.0
SQLAlchemy==1.4.48
Flask-SQLAlchemy==2.5.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
jsonschema==4.19.1
python-dateutil==2.8.2
celery==5.3.2
redis==4.6.0